SmartWeather/requirements.txt
flask==3.1.2
flask-cors==6.0.1
jinja2==3.1.6
pyyaml==6.0.3
jsonpickle==4.1.1
ollama==0.6.0
mcp==1.1.2
openai==2.7.2
requests==2.32.3
